pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["evg671ZXC <evgmine671@gmail.com>"]
readme = "README.md"
packages = [{ include = "page_analyzer" }, ]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flake8 = "^7.1.0"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
validators = "^0.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
